# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Acestream Scraper API",
    "description": "API for scraping and managing Acestream channels",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/playlists/m3u": {
      "get": {
        "tags": ["playlists"],
        "summary": "Get M3U Playlist",
        "description": "Generate M3U playlist with specified filters\n\n- **search**: Optional search term for channel names\n- **group**: Optional specific group to filter by\n- **only_online**: Whether to include only online channels (default: True)\n- **include_groups**: Comma-separated list of groups to include\n- **exclude_groups**: Comma-separated list of groups to exclude",
        "operationId": "get_m3u_playlist_api_v1_playlists_m3u_get",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Search"
            }
          },
          {
            "name": "group",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Group"
            }
          },
          {
            "name": "only_online",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Only Online"
            }
          },
          {
            "name": "include_groups",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Include Groups"
            }
          },
          {
            "name": "exclude_groups",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Exclude Groups"
            }
          },
          {
            "name": "base_url",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Base Url"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Format"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "text/plain": { "schema": { "type": "string" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/playlists/playlists/m3u": {
      "get": {
        "tags": ["playlists"],
        "summary": "Get M3U Playlist Compat",
        "description": "Compatibility route for M3U playlist generation\n\nThis route exists to support legacy URLs. It behaves identically to the\n/m3u endpoint.",
        "operationId": "get_m3u_playlist_compat_api_v1_playlists_playlists_m3u_get",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Search"
            }
          },
          {
            "name": "group",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Group"
            }
          },
          {
            "name": "only_online",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Only Online"
            }
          },
          {
            "name": "include_groups",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Include Groups"
            }
          },
          {
            "name": "exclude_groups",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Exclude Groups"
            }
          },
          {
            "name": "base_url",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Base Url"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Format"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "text/plain": { "schema": { "type": "string" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/search": {
      "get": {
        "tags": ["search", "search"],
        "summary": "Search Channels",
        "description": "Search for Acestream channels via engine API",
        "operationId": "search_channels_api_v1_search_get",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Search query string (optional)",
              "default": "",
              "title": "Query"
            },
            "description": "Search query string (optional)"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number (default: 1, must be >= 1)",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number (default: 1, must be >= 1)"
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Results per page (default: 10, must be >= 1, alias: per_page)",
              "default": 10,
              "title": "Per Page"
            },
            "description": "Results per page (default: 10, must be >= 1, alias: per_page)"
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "description": "Filter by category (optional)",
              "title": "Category"
            },
            "description": "Filter by category (optional)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SearchResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/search/add": {
      "post": {
        "tags": ["search", "search"],
        "summary": "Add Channel",
        "description": "Add a channel from search results to the database (delegated to service)",
        "operationId": "add_channel_api_v1_search_add_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AddChannelRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/search/add_multiple": {
      "post": {
        "tags": ["search", "search"],
        "summary": "Add Multiple Channels",
        "description": "Add multiple channels from search results to the database (delegated to service)",
        "operationId": "add_multiple_channels_api_v1_search_add_multiple_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AddMultipleRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddChannelRequest": {
        "properties": {
          "id": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Id",
            "description": "Acestream channel ID (optional, will be generated if not provided)"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Channel name"
          },
          "url": {
            "type": "string",
            "title": "Url",
            "description": "Channel URL"
          },
          "category": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Category",
            "description": "Channel category"
          },
          "description": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Description",
            "description": "Channel description"
          },
          "logo": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Logo",
            "description": "Channel logo URL"
          },
          "tv_channel_id": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Tv Channel Id",
            "description": "TV channel ID for association"
          }
        },
        "type": "object",
        "required": ["name", "url"],
        "title": "AddChannelRequest"
      },
      "AddMultipleRequest": {
        "properties": {
          "channels": {
            "items": { "$ref": "#/components/schemas/AddChannelRequest" },
            "type": "array",
            "title": "Channels",
            "description": "List of channels to add"
          }
        },
        "type": "object",
        "required": ["channels"],
        "title": "AddMultipleRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SearchPagination": {
        "properties": {
          "page": {
            "type": "integer",
            "title": "Page",
            "description": "Current page number"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size",
            "description": "Results per page"
          },
          "total_results": {
            "type": "integer",
            "title": "Total Results",
            "description": "Total number of results"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages",
            "description": "Total number of pages"
          }
        },
        "type": "object",
        "required": ["page", "page_size", "total_results", "total_pages"],
        "title": "SearchPagination"
      },
      "SearchResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Success status of the request"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Message describing the result"
          },
          "results": {
            "items": { "$ref": "#/components/schemas/SearchResultItem" },
            "type": "array",
            "title": "Results",
            "description": "Search results"
          },
          "pagination": {
            "$ref": "#/components/schemas/SearchPagination",
            "description": "Pagination information"
          }
        },
        "type": "object",
        "required": ["success", "message", "pagination"],
        "title": "SearchResponse"
      },
      "SearchResultItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Acestream channel ID (infohash)"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Channel name"
          },
          "bitrate": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Bitrate",
            "description": "Channel bitrate"
          },
          "categories": {
            "items": { "type": "string" },
            "type": "array",
            "title": "Categories",
            "description": "Categories of the channel"
          }
        },
        "type": "object",
        "required": ["id", "name"],
        "title": "SearchResultItem"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    }
  }
}